llm_rag_translate_pdf_page/pyproject.toml
[tool.poetry]
name = "llm-rag-translate-pdf-page"
version = "0.1.0"
description = ""
authors = ["ted park <itstedpark@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
gradio = "^5.5.0"
langchain-openai = "^0.2.5"
langchain = "^0.3.7"
langchain-community = "^0.3.5"
pypdf = "^5.1.0"
pymupdf = "^1.24.13"
transformers = "^4.46.1"
streamlit = "^1.40.0"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
